Schematic updated
Added a several new labels on the wiring diagram and made a section for the recovery altimeters, also made the e-matches necessary to light the black powder. Added recurring labels to wires on the right side and marked many ports with no X's to mean that they will not be connected to anything during wiring.